encode: allow for a custom `special` object
By default the `encode()` function will encode all non-safe
HTML characters.

But in some contexts, you don't want to encode *all* of these.
i.e. consider encoding a TextNode, the `"` and `'` chars are safe
in that context, but `<` `>` and `&` should still be encoded.

This allows for setting an explicit object to use.